requirements.txt
# Dependencies for Ark OS
requests>=2.28.0
pyserial>=3.5
